This commit resolves an issue where usage hours set to fractional values (e.g., 0.5) were incorrectly displayed as 0. The implementation now ensures that fractional values are accurately shown to enhance user experience and calculation accuracy.

+++ app/DTOs/ApplianceDTO.php
@@ -17,7 +17,7 @@ class ApplianceDTO implements DTO
         $this->id = $data['id'] ?? null;
         $this->name = $data['name'] ?? '';
         $this->powerRating = (int) ($data['powerRating'] ?? 0);
-        $this->usageHours = (int) ($data['usageHours'] ?? 0);
+        $this->usageHours = (float) ($data['usageHours'] ?? 0);
         $this->units = (int) ($data['units'] ?? 1);
         $this->createdAt = $data['createdAt'] ?? '';
         $this->updatedAt = $data['createdAt'] ?? '';
